externalize verb ending data; rename spanish utils to expected case format
