+++ data/verb-endings.json
@@ -1,0 +1,1 @@
+{"verbEndings":["ar","er","ir","arse","erse","irse","ár","ér","ír","árse","érse","írse"],"irEndings":{"1pre":"o","2pre":"es","3pre":"e","4pre":"imos","5pre":"ís","6pre":"en","1pas":"í","2pas":"iste","3pas":"ió","4pas":"imos","5pas":"isteis","6pas":"ieron","1fut":"iré","2fut":"irás","3fut":"irá","4fut":"iremos","5fut":"iréis","6fut":"irán","1cop":"ía","2cop":"ías","3cop":"ía","4cop":"íamos","5cop":"íais","6cop":"ían","1pos":"iría","2pos":"irías","3pos":"iría","4pos":"iríamos","5pos":"iríais","6pos":"irían","1pres":"a","2pres":"as","3pres":"a","4pres":"amos","5pres":"áis","6pres":"an","1pass":"iera","2pass":"ieras","3pass":"iera","4pass":"iéramos","5pass":"ierais","6pass":"ieran","1passb":"iese","2passb":"ieses","3passb":"iese","4passb":"iésemos","5passb":"ieseis","6passb":"iesen","1futs":"iere","2futs":"ieres","3futs":"iere","4futs":"iéremos","5futs":"iereis","6futs":"ieren","2imp":"e","3imp":"a","4imp":"amos","5imp":"id","6imp":"an","gerundio":"iendo","participio":"ido"},"erEndings":{"1pre":"o","2pre":"es","3pre":"e","4pre":"emos","5pre":"éis","6pre":"en","1pas":"í","2pas":"iste","3pas":"ió","4pas":"imos","5pas":"isteis","6pas":"ieron","1fut":"eré","2fut":"erás","3fut":"erá","4fut":"eremos","5fut":"eréis","6fut":"erán","1cop":"ía","2cop":"ías","3cop":"ía","4cop":"íamos","5cop":"íais","6cop":"ían","1pos":"ería","2pos":"erías","3pos":"ería","4pos":"eríamos","5pos":"eríais","6pos":"erían","1pres":"a","2pres":"as","3pres":"a","4pres":"amos","5pres":"áis","6pres":"an","1pass":"iera","2pass":"ieras","3pass":"iera","4pass":"iéramos","5pass":"ierais","6pass":"ieran","1passb":"iese","2passb":"ieses","3passb":"iese","4passb":"iésemos","5passb":"ieseis","6passb":"iesen","1futs":"iere","2futs":"ieres","3futs":"iere","4futs":"iéremos","5futs":"iereis","6futs":"ieren","2imp":"e","3imp":"a","4imp":"amos","5imp":"ed","6imp":"an","gerundio":"iendo","participio":"ido"},"arEndings":{"1pre":"o","2pre":"as","3pre":"a","4pre":"amos","5pre":"áis","6pre":"an","1pas":"é","2pas":"aste","3pas":"ó","4pas":"amos","5pas":"steis","6pas":"aron","1fut":"aré","2fut":"arás","3fut":"ará","4fut":"aremos","5fut":"aréis","6fut":"arán","1cop":"aba","2cop":"abas","3cop":"aba","4cop":"ábamos","5cop":"abais","6cop":"aban","1pos":"aría","2pos":"arías","3pos":"aría","4pos":"aríamos","5pos":"aríais","6pos":"arían","1pres":"e","2pres":"es","3pres":"e","4pres":"emos","5pres":"éis","6pres":"en","1pass":"ara","2pass":"aras","3pass":"ara","4pass":"áramos","5pass":"arais","6pass":"aran","1passb":"ase","2passb":"ases","3passb":"ase","4passb":"ásemos","5passb":"aseis","6passb":"asen","1futs":"are","2futs":"ares","3futs":"are","4futs":"áremos","5futs":"areis","6futs":"aren","2imp":"a","3imp":"e","4imp":"emos","5imp":"ad","6imp":"en","gerundio":"ando","participio":"ado"},"zarEndings":{"1pas":"cé","1pres":"ce","2pres":"ces","3pres":"ce","4pres":"cemos","5pres":"céis","6pres":"cen","3imp":"ce","4imp":"cemos","6imp":"cen"},"carEndings":{"1pas":"qué","1pres":"que","2pres":"ques","3pres":"que","4pres":"quemos","5pres":"quéis","6pres":"quen","3imp":"que","4imp":"quemos","6imp":"quen"},"garEndings":{"1pas":"ué","1pres":"ue","2pres":"ues","3pres":"ue","4pres":"uemos","5pres":"uéis","6pres":"uen","3imp":"ue","4imp":"uemos","6imp":"uen"}}
