+++ services/diff.js
@@ -1,33 +1,33 @@
-var _ = require('lodash');
-var googlediff = require('googlediff');
-var seedrandom = require('seedrandom');
+const _ = require('lodash');
+const GoogleDiff = require('googlediff');
+const seedrandom = require('seedrandom');
 
-googlediff = new googlediff();
+const gdiff = new GoogleDiff();
 
-var ITERATIONS = 10000;
-var ALPHABET = 'GATTACA';
-var LENGTH = 100;
+const ITERATIONS = 10000;
+const ALPHABET = 'GATTACA';
+const LENGTH = 100;
 
-var seed = Math.floor(Math.random() * 10000);
-var random = seedrandom(seed);
+const seed = Math.floor(Math.random() * 10000);
+const random = seedrandom(seed);
 
 console.log('Running computing ' + ITERATIONS + ' diffs with seed ' + seed + '...');
 
 console.log('Generating strings...');
-var strings = [];
-for(var i = 0; i <= ITERATIONS; ++i) {
-  var chars = [];
-  for(var l = 0; l < LENGTH; ++l) {
-    var letter = ALPHABET.substr(Math.floor(random() * ALPHABET.length), 1);
+const strings = [];
+for(let i = 0; i <= ITERATIONS; ++i) {
+  const chars = [];
+  for(let l = 0; l < LENGTH; ++l) {
+    const letter = ALPHABET.substr(Math.floor(random() * ALPHABET.length), 1);
     chars.push(letter);
   }
   strings.push(chars.join(''));
 }
 
 console.log('Running tests *without* cursor information...');
-for(var i = 0; i < ITERATIONS; ++i) {
-  var result = diff(strings[i], strings[i+1]);
-  var expected = googlediff.diff_main(strings[i], strings[i+1]);
+for(let i = 0; i < ITERATIONS; ++i) {
+  const result = diff(strings[i], strings[i+1]);
+  const expected = gdiff.diff_main(strings[i], strings[i+1]);
   if (!_.isEqual(result, expected)) {
     console.log('Expected', expected);
     console.log('Result', result);
@@ -36,11 +36,11 @@ for(var i = 0; i < ITERATIONS; ++i) {
 }
 
 console.log('Running tests *with* cursor information');
-for(var i = 0; i < ITERATIONS; ++i) {
-  var cursor_pos = Math.floor(random() * strings[i].length + 1);
-  var diffs = diff(strings[i], strings[i+1], cursor_pos);
-  var patch = googlediff.patch_make(strings[i], strings[i+1], diffs);
-  var expected = googlediff.patch_apply(patch, strings[i])[0];
+for(let i = 0; i < ITERATIONS; ++i) {
+  const cursor_pos = Math.floor(random() * strings[i].length + 1);
+  const diffs = diff(strings[i], strings[i+1], cursor_pos);
+  const patch = gdiff.patch_make(strings[i], strings[i+1], diffs);
+  const expected = gdiff.patch_apply(patch, strings[i])[0];
   if (expected !== strings[i+1]) {
     console.log('Expected', expected);
     console.log('Result', strings[i+1]);
